+++ semcore/wizard/CHANGELOG.md
@@ -2,7 +2,13 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [1.0.0-0] - unreleased
+## [1.0.0] - 2022-06-29
+
+### Changed
+
+- Changed api
+
+## [1.0.0-0] - 2022-06-01
 
 ### Added